feat: allow fractional key pitch in panel geometry
Holding both margins on an axis fixed while adjusting one of them puts
the difference into the gaps, which rarely comes out whole. Pitch is now
an f32 on both axes; plugins declaring integers are unaffected.

+++ src-tauri/src/shared.rs
@@ -73,8 +73,10 @@ pub struct PanelInfo {
 	pub keys_x: u16,
 	pub keys_y: u16,
 	pub key_size: u16,
-	pub pitch_x: u16,
-	pub pitch_y: u16,
+	/// Pitch may be fractional: holding both margins fixed while moving one
+	/// distributes the difference across the gaps.
+	pub pitch_x: f32,
+	pub pitch_y: f32,
 }
 
 pub static DEVICES: LazyLock<DashMap<String, DeviceInfo>> = LazyLock::new(DashMap::new);
